Separated SSE and stdio transports into their own files
