+++ Cargo.toml
@@ -38,10 +38,10 @@ async-recursion = "1.1.1"
 name = "server"
 path = "bin/server.rs"
 
-
 [[bin]]
 name = "client"
 path = "bin/client.rs"
+
 [lib]
 name = "mcp_rs"
 path = "src/lib.rs"
